feat(EPIC-001): widen TokenCreationPayloadV1.initial_supply to u128
Closes the last EPIC-001 decimals-unification straggler. lib-client's
build_create_token_tx already accepted u128 and silently downcast at the
wire boundary (`initial_supply as u64 // TODO: widen TokenCreationPayloadV1
to u128`, PR #381bd0ef Apr 15). The chain-side payload was deliberately
left at u64 "for bincode compat" — but that was a workaround, not a
final state, and it capped every token created via this path at
~18.4 BUBL display @ 18 decimals (1.84e19 atoms). CBE itself isn't
affected because it ships through `BondingCurveToken::deploy` at genesis
(u128 native), not through `TokenCreation`.

Widens `TokenCreationPayloadV1.initial_supply` and
`split_initial_supply` to `u128` end-to-end:

- `lib-blockchain/src/transaction/token_creation.rs`: struct field +
  splitter return type. Splitter uses `checked_mul` (with
  `treasury_allocation_bps <= 10_000` enforced by `validate`) instead of
  `u64::try_from().expect()`, so a bogus payload that bypasses
  `validate` cannot panic at split time.
- `lib-blockchain/src/blockchain/contracts.rs` + `execution/executor.rs`:
  drop the `as u128` mint downcasts that are now identity casts.
- Test helpers in `executor.rs` and `replay_contract_execution_tests.rs`
  take `u128 supply`. Test literals (`100`, `1_000`, `1_000_000`) infer
  from the field type — no literal changes needed.
- `lib-client/src/token_tx.rs`: remove the `as u64` boundary downcast.
  Public API was already u128.
- `zhtp-cli/src/argument_parsing.rs` + `commands/token.rs`: `--supply`
  arg widened to u128 in clap + handler.
- `zhtp/src/api/handlers/token/mod.rs`: serialise u128 fields
  (`initial_supply`, `creator_allocation`, `treasury_allocation`) as
  decimal strings in the JSON response, matching the pattern already
  established by earlier u128/JSON serialization fixes (commit 73fb0eab,
  9f45d39a) — serde_json cannot represent u128 numerically.

Wire-format / deploy considerations:

- Tx-resident only — `TokenCreationPayloadV1` is never persisted in
  sled. No PR #2623-class EOF-on-old-blob risk.
- No `TokenCreation` transaction has ever been processed on chain
  (verified by 30-day journal sweep on g1). No mempool replay, no
  historical decode break.
- Old lib-client clients in the field that still encode `u64`
  initial_supply will fail tx decode on the new server — clean error,
  no crash. Practical impact: zero, since no production user has ever
  submitted a `TokenCreation` tx.

Tests: lib-blockchain 1863/1863, lib-client 61/61, plus 17/17 token_creation
suite tests (canonical happy path, fee-zero/fee-above-canonical bounds,
self-treasury rejection, symbol case-insensitive dedup, duplicate
rejection, legacy-flow self-treasury rejection). `cargo build -p
lib-blockchain -p lib-client -p zhtp-cli -p zhtp` clean.

+++ lib-blockchain/src/blockchain/contracts.rs
@@ -681,16 +681,16 @@ impl Blockchain {
                     } else {
                         payload.decimals
                     };
-                    token.max_supply = payload.initial_supply as u128;
+                    token.max_supply = payload.initial_supply;
                     token
-                        .mint(&creator, creator_allocation as u128)
+                        .mint(&creator, creator_allocation)
                         .map_err(|e| anyhow::anyhow!("TokenCreation mint failed: {}", e))?;
                     let treasury_pk = lib_crypto::types::keys::PublicKey {
                         dilithium_pk: [0u8; 2592],
                         kyber_pk: [0u8; 1568],
                         key_id: payload.treasury_recipient,
                     };
-                    token.mint(&treasury_pk, treasury_allocation as u128).map_err(|e| {
+                    token.mint(&treasury_pk, treasury_allocation).map_err(|e| {
                         anyhow::anyhow!("TokenCreation treasury mint failed: {}", e)
                     })?;
